Git API Code Validation
With lots of testing, all this code is now fully functional!
I have refactored the commits and contributors of the repository to be stored in the larger datastructure as this makes it even more database friendly. Am looking to revise filechanges in the near future to include additional information.

TODO: Flow Diagram of the inner machinations of this codebase