Feat/refactor heatmap (#47)
* feat: add HeatmapGraph component to AnalyticsView
and refactor to use ContributionEntry type

* fix: update heatmap axis labeling to always have 7 columns

* fix: clean up NavBar component to fix console error

Nesting <a> tags aren't allowed which was happening when a <Link> is nested within a <NavigationMenuLink> which renders as a <a>

* fix: simplify y-axis labels in HeatmapGraph component for Weeks mode

* fix: adjust heatmap data generation and update color levels for better visualization

* fix: update heatmap description and adjust color level for improved clarity

* fix: correct import statement for HeatmapGraph component

* fix: update color levels in HeatmapGraph for improved visualization

* fix: update heatmap data calculation and improve description for clarity

Previously, the heatmap color intensity for each cell was normalized based on each user's own highest activity, highlighting individual patterns. With this change, color intensity is now normalized by the highest contribution made by any user in each time bucket. This allows for direct comparison of user activity within the same period, making it easier to identify top contributors for each time segment. Tooltip and title have been updated to reflect this new behavior.

* fix: update heatmap data generation to use date range from state and adjust default days back

* fix: remove generateHeatmapData function and replace with static heatmap data for contributor activity

* fix: update date in contributor activity heatmap and refine year mode logic in HeatmapGraph

* fix: changing spelling error in heatmap graph import.

* fix: fix broken navbar

* feat: update HeatmapGraph to use HeatMapData type and handle empty data case

---------

Co-authored-by: Milni Abeysekara <mabe0012@student.monash.edu>